Code cleaning, refactor ("html" and "text" functions now being exported in a own "utils.js" file and used in [...nextauth].js, and useContext logic removed regarding the session info (AdminTobBar component).
