+++ components/Admin/Nav/Avatar/Avatar.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import imgAvatar from "../../../../pages/assets/images/image-avatar.jpg";
 import styles from "./Avatar.module.css";
 import { useContext } from "react";
 import IssueContext from "@/utils/IssueContext";
@@ -8,7 +7,6 @@ export default function Avatar() {
   const context = useContext(IssueContext);
   return (
     <div className={styles.main}>
-      {/* <img className={styles.image} src={imgAvatar.src} alt="admin avatar" /> */}
       <img
         className={styles.image}
         src={context?.signedInAs?.user?.image}
