+++ pages/_app.js
@@ -10,11 +10,8 @@ export default function App({
   Component,
   pageProps: { session, ...pageProps },
 }) {
-  //console.log("session", session);
-
   const [loading, setLoading] = useState(false);
   const [openModal, setOpenModal] = useState(false);
-  const [signedInAs, setSignedInAs] = useState(null);
 
   return (
     <SessionProvider session={session}>
@@ -24,8 +21,6 @@ export default function App({
           setLoading,
           openModal,
           setOpenModal,
-          signedInAs,
-          setSignedInAs,
         }}
       >
         {" "}
