+++ pages/api/auth/[...nextauth].js
@@ -2,9 +2,8 @@ import NextAuth from "next-auth";
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
 import EmailProvider from "next-auth/providers/email";
 import { createTransport } from "nodemailer";
-import GithubProvider from "next-auth/providers/github";
-
 import { PrismaClient } from "@prisma/client";
+import { text, html } from "./utils";
 
 const prisma = new PrismaClient();
 
@@ -48,55 +47,8 @@ export const authOptions = {
           throw new Error(`Email(s) (${failed.join(", ")}) could not be sent`);
         }
 
-        function html({ url, host, theme }) {
-          const escapedHost = host.replace(/\./g, "&#8203;.");
-
-          const brandColor = theme.brandColor || "#346df1";
-          const color = {
-            background: "#f9f9f9",
-            text: "#444",
-            mainBackground: "#fff",
-            buttonBackground: brandColor,
-            buttonBorder: brandColor,
-            buttonText: theme.buttonText || "#fff",
-          };
-
-          return `
-          <body style="background: ${color.background};">
-            <table width="100%" border="0" cellspacing="20" cellpadding="0"
-              style="background: ${color.mainBackground}; max-width: 600px; margin: auto; border-radius: 10px;">
-              <tr>
-                <td align="center"
-                  style="padding: 10px 0px; font-size: 22px; font-family: Helvetica, Arial, sans-serif; color: ${color.text};">
-                  Sign in to <strong>${escapedHost}</strong>
-                </td>
-              </tr>
-              <tr>
-                <td align="center" style="padding: 20px 0;">
-                  <table border="0" cellspacing="0" cellpadding="0">
-                    <tr>
-                      <td align="center" style="border-radius: 5px;" bgcolor="${color.buttonBackground}"><a href="${url}"
-                          target="_blank"
-                          style="font-size: 18px; font-family: Helvetica, Arial, sans-serif; color: ${color.buttonText}; text-decoration: none; border-radius: 5px; padding: 10px 20px; border: 1px solid ${color.buttonBorder}; display: inline-block; font-weight: bold;">Sign
-                          in</a></td>
-                    </tr>
-                  </table>
-                </td>
-              </tr>
-              <tr>
-                <td align="center"
-                  style="padding: 0px 0px 10px 0px; font-size: 16px; line-height: 22px; font-family: Helvetica, Arial, sans-serif; color: ${color.text};">
-                  If you did not request this email ::::::: you can safely ignore it.
-                </td>
-              </tr>
-            </table>
-          </body>
-          `;
-        }
-
-        function text({ url, host }) {
-          return `Sign in to ${host}\n${url}\n\n`;
-        }
+        html({ url, host, theme });
+        text({ url, host });
       },
     }),
     {},
@@ -111,7 +63,10 @@ export const authOptions = {
 };
 export default NextAuth(authOptions);
 
-//alternativelly Providers.GitHub
+// For creating new users, we can enable Github provider like this:
+
+// import GithubProvider from "next-auth/providers/github";
+// (...)
 // GithubProvider({
 //   clientId: process.env.GITHUB_ID,
 //   clientSecret: process.env.GITHUB_SECRET,
